src/inverse_problem_with_diffusion_prior/requirements.txt
absl-py==1.4.0
cached-property==1.5.2
certifi==2022.12.7
charset-normalizer==3.0.1
chex==0.1.6
contourpy==1.0.7
cycler==0.11.0
dm-tree==0.1.8
etils==1.0.0
filelock==3.9.0
flax==0.6.6
fonttools==4.38.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
jax==0.4.4
jaxlib==0.4.4+cuda11.cudnn82
kiwisolver==1.4.4
markdown-it-py==2.2.0
matplotlib==3.7.0
mdurl==0.1.2
msgpack==1.0.4
multipledispatch==0.6.0
numpy==1.24.2
numpyro==0.11.0
opt-einsum==3.3.0
optax==0.1.4
orbax==0.1.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pkg_resources==0.0.0
psutil==5.9.4
Pygments==2.14.0
pyparsing==3.0.9
pyro-api==0.1.2
pyro-ppl==1.8.4
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rich==13.3.1
scipy==1.10.1
six==1.16.0
tensorstore==0.1.33
toolz==0.12.0
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
zipp==3.15.0
